# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FreshLux Carpet Cleaning UK API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: string;
  name: string;
  slug: string;
  description: string;
  /** Starting price in GBP pence (divide by 100 for pounds) */
  startingFrom: number;
  icon: string;
  features?: string[];
}

export interface ServiceItem {
  id: string;
  name: string;
  category: string;
  /** Price in GBP (e.g. 45.00) */
  priceGbp: number;
  /** e.g. "per item", "per m²" */
  unit: string;
  /** @nullable */
  description?: string | null;
}

export interface QuoteLineItem {
  itemId: string;
  quantity: number;
}

export interface QuoteInput {
  serviceId: string;
  items: QuoteLineItem[];
  addons?: string[];
  /** @nullable */
  couponCode?: string | null;
}

export interface QuoteLine {
  name: string;
  quantity: number;
  unitPrice: number;
  lineTotal: number;
}

export interface QuoteResult {
  subtotal: number;
  discount: number;
  total: number;
  minimumApplied: boolean;
  savedAmount?: number;
  lines: QuoteLine[];
}

export interface CouponValidateInput {
  code: string;
}

export type CouponResultDiscountType = typeof CouponResultDiscountType[keyof typeof CouponResultDiscountType];


export const CouponResultDiscountType = {
  percentage: 'percentage',
  fixed: 'fixed',
} as const;

export interface CouponResult {
  code: string;
  discountType: CouponResultDiscountType;
  discountValue: number;
  description: string;
}

export type BookingInputPropertyType = typeof BookingInputPropertyType[keyof typeof BookingInputPropertyType];


export const BookingInputPropertyType = {
  flat: 'flat',
  house: 'house',
  office: 'office',
} as const;

export interface BookingInput {
  serviceId: string;
  items: QuoteLineItem[];
  addons?: string[];
  /** ISO date string YYYY-MM-DD */
  date: string;
  /** e.g. "09:00-11:00" */
  timeSlot: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  postcode: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  couponCode?: string | null;
  propertyType?: BookingInputPropertyType;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  bookingNumber: string;
  serviceId: string;
  serviceName: string;
  date: string;
  timeSlot: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  postcode: string;
  /** @nullable */
  notes?: string | null;
  status: BookingStatus;
  totalGbp: number;
  subtotalGbp?: number;
  discountGbp?: number;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  propertyType?: string | null;
  items?: QuoteLine[];
  createdAt: string;
}

export interface AvailableSlot {
  date: string;
  slots: string[];
}

export interface ContactInput {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  message: string;
  /** @nullable */
  subject?: string | null;
}

export interface ContactResult {
  success: boolean;
  message: string;
}

export type GetAvailabilityParams = {
/**
 * ISO date string (YYYY-MM-DD)
 */
from: string;
/**
 * ISO date string (YYYY-MM-DD)
 */
to: string;
};