# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FreshLux Carpet Cleaning UK API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';
import type { QuoteLine } from './quoteLine';

export interface Booking {
  id: number;
  bookingNumber: string;
  serviceId: string;
  serviceName: string;
  date: string;
  timeSlot: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  postcode: string;
  /** @nullable */
  notes?: string | null;
  status: BookingStatus;
  totalGbp: number;
  subtotalGbp?: number;
  discountGbp?: number;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  propertyType?: string | null;
  items?: QuoteLine[];
  createdAt: string;
}