# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FreshLux Carpet Cleaning UK API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingInputPropertyType } from './bookingInputPropertyType';
import type { QuoteLineItem } from './quoteLineItem';

export interface BookingInput {
  serviceId: string;
  items: QuoteLineItem[];
  addons?: string[];
  /** ISO date string YYYY-MM-DD */
  date: string;
  /** e.g. "09:00-11:00" */
  timeSlot: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  postcode: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  couponCode?: string | null;
  propertyType?: BookingInputPropertyType;
}