# lib/api-zod/src/generated/types/couponResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FreshLux Carpet Cleaning UK API
 * OpenAPI spec version: 0.1.0
 */
import type { CouponResultDiscountType } from './couponResultDiscountType';

export interface CouponResult {
  code: string;
  discountType: CouponResultDiscountType;
  discountValue: number;
  description: string;
}