# lib/api-zod/src/generated/types/quoteLine.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FreshLux Carpet Cleaning UK API
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteLine {
  name: string;
  quantity: number;
  unitPrice: number;
  lineTotal: number;
}