# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FreshLux Carpet Cleaning UK API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: string;
  name: string;
  slug: string;
  description: string;
  /** Starting price in GBP pence (divide by 100 for pounds) */
  startingFrom: number;
  icon: string;
  features?: string[];
}